React Versions changes and remove typescript error
