+++ next.config.js
@@ -1,0 +1,9 @@
+/**
+ * @type {import('next').NextConfig}
+ */
+const nextConfig = {
+  reactStrictMode: true, // Recommended for catching issues early
+ 
+};
+
+module.exports = nextConfig;
