+++ package.json
@@ -11,9 +11,9 @@
   "dependencies": {
     "@types/react-modal": "^3.16.3",
     "firebase": "^11.0.2",
-    "next": "15.0.3",
-    "react": "19.0.0-rc-66855b96-20241106",
-    "react-dom": "19.0.0-rc-66855b96-20241106",
+    "next": "^14.2.21",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
     "react-loader-spinner": "^6.1.6",
     "react-modal": "^3.16.1",
